+++ Negocio.Model/Utilidades/Municipio.cs
@@ -2,9 +2,9 @@
 {
     public class Municipio
     {
-        public int Codigo { get; set; }
+        public decimal Codigo { get; set; }
         public string Codigomunicio { get; set; }
         public string Nombre { get; set; }
-        public int Codigodepto { get; set; }
+        public decimal Codigodepto { get; set; }
     }
 }
